renamed console_gui package to console
