+++ src/commands/WestCommand.java
@@ -1,11 +1,6 @@
 package commands;
 
-/**
- * A command to move the user in a given direction.
- * 
- * @author Zachary Chandler
- */
-import console_gui.User;
+import console.User;
 import model.Direction;
 
 public class WestCommand extends MoveCommand {
