+++ src/console/Console.java
@@ -1,4 +1,4 @@
-package console_gui;
+package console;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
