+++ src/console/PlayerManager.java
@@ -1,4 +1,4 @@
-package console_gui;
+package console;
 
 import java.util.Collections;
 import java.util.LinkedList;
